captain.lua: Change Versus mode text color to be readable
It has been magenta on aqua (water tiles) until now. It was not a
pleasant color combination to read even with good color sight. Changed
the vesus color to orange, it's supposed to be better for color
blindness palettes I've found.

color_presets.lua is mixed type between 0-255 and 0-1.0 for now...