Add a new research info window (disabled by default) (#490)
* Add a new research info window (disabled by default)

This is meant to replace the white-flask that allows spectators to
view research progress during captains games. Specifically, the
current implementation bothers me because it actually moves players
between forces, which I believe introduces many possible
races/confusion/bugs. I also hope that this implementation might be
more useful, as it shows a comparison of both teams at the same time.

I believe that bugs with the existing implementation led to the
decision to only enable the white-flask during captain's games. I hope
that this new implementation can be enabled in all games.

For now, it is disabled by default, but can be enabled by setting
  global.bb_show_research_info = "always"
or
  global.bb_show_research_info = "spec"
to enable it, and setting
  global.bb_show_research_info = nil
to disable it

I find UI code frustrating to write, because I am accutely aware of
millions of ways in which the final UI is not perfect, but it takes a
long time and many iterations to make changes, and at some ponit I
have to decide "good enough for now".