Changes to the pathfinder settings (#502)
Changes to the pathfinder to:

Make the short path cache much bigger and increase the minimal length of a path to be cached into the short path cache to make sure that really short paths and thus fast to compute won't fill the cache and override longer paths. I found the value of 12 by testing some and taking the value that filled the cache to a little less than 100% when waves attack walls. Should reduce lag.

Increase the minimal length of a path to be cached into the long path cache to reserve this cache for wave paths. With this setting set to 60, I have seen some situation where a group of biters will be distracted by something (and each of them will calculate a 1 path to that thing) that is more than 60 tiles away and thus completely override all the 50 wave paths cached. This happens when a player attacks a spawner to threat farm and that all biters near that spawner attack the player, which can be more than 60 tiles away. Should make the pathfinder faster to compute wave paths.

Reduce the entity collision penalty, should reduce lag and solve the problem of AFK waves we often see.

Reduce the path finder resolution to -1, should reduce lag but make the biters a little more dumb when attacking checkerboard-like walls. I have set the modifier to -1 but we may want reduce it even more later if the pathfinder still causes lag. Note: this modifier has a really predictable impact of performance and should multiple the amount of work of the pathfinder by 2^2n.

Co-authored-by: ness056 <ness056@ness056>