Remove "tech-spy" function from captains games
It has been replaced by the new research UI that shows information
about both teams doing research.

Specifically, this function is very dangerous, because it actually
swaps players to different forces in order to look at the research UI.
If something happens concurrently, like a player being chosen to join
a team, or a referee moving a player, or the player actually joining a
team normally (not normally possible in a captains game)), then this
can corrupt state. Also, it is confusing because the spectator using
this button will show up in the list of players for whatever force
they are looking at.  Also, the number of "active players" is used for
things like threat-generation calculations, so, for instance, in a 2v2
tournament situation, if a player is looking at technology at the same
time that someone sends, it will actually generate additional threat,
which seems like a bad idea.