Admin option: MoveToSpawn (#500)
This can be used to teleport a player to spawn position.  This is
functionally fairly similar to "Kill", except that they can keep their
inventory and it is possibly less offensive.

Additionally this makes the various teleport options work even if
players are driving (it just removes them from their vehicle first,
the vehicle will be left behind).