Stop always broadcasting admin chat messages (#510)
This was voted on by the admins and universally agreed to be confusing
behavior. It is confusing for admins because they have to remember to
use /spectator-chat. It is confusing for non-admins because they see
chat messages that are not /shouts from other players, and assume that
perhaps their chat messages work the same way.