Initial/experimental support for @ping in chat
Right now, if any chat message you can see contains "@username" or "@
username", or if you are the target of a wisper, then the following
effects happen:
1) if you are at ~100% health, you take 1 damage
2) a "new objective" sound is played at 0.6 volume
3) a popup appears at the top of the screen that contains the message
   for 10 seconds

I made these on things that were hopefully not too annoying, but also
hard to miss.

Right now, this needs to be manually enabled under Fish -> Config for
each user that wants to see the ping effects.

I'm hoping to eventually enable something like this by default
(although people could still disable it), but I need feedback about
what the right effect is first (and if we need rate limiting or
something first).