Captains more improvements (#519)
* Disable group-picks after the initial round of picking

It is too hard for picking rounds to be balanced otherwise.

* captains join UI: cleanup

Now buttons are much less likely to move around (I just disable them
rather than turn them invisible). Also the "I want to be captain"
button is always visible and distinct from the "I want to play"
button.

* Captains event: allow referees to remove captain volunteers at any time

This should allow it to be clearer, for longer, who is playing in the
game. This also simplifies code.

* Captains: do not list people as both "disconnected" and "spectating"