Captains games: add randomized-fair picking mode (#524)
This takes, as an input, a coarse-grained view of how good players
are, and then makes two random teams that are mostly-fair. At the
moment, it relies on the referee to build this
approximate-stack-rank. However, I plan on allowing captains to also
take part in it as well.

This also allows referees to configure more things after the special
has started (but before captains are finalized).

I spent way too long figuring out that "freeze_players" made it so
that people could not interact with switches in the UI.

I also added a test-only command that is very very useful for me when
testing this functionality.

For now, we only do the randomized-fair picking for the first picking round

Specifically, the current poll/popups for other rounds are valuable to
captains so that they know who is on their team.