Partially revert pathfinder settings change (#502) (#527)
This is primarily motivated because I believe that the high value of
short_cache_size was drastically hurting performance in some
situations, based on
https://forums.factorio.com/viewtopic.php?p=610324#p610324

So, the main thing that I wanted to do here is to move things closer
to their default values, and closer to the values that they were
before #502.  Note that I am leaving the lower-pathfinder-resolution,
as I assume that that is still a big speedup for the pathfinder in
general (even though the downsides are substantial).

I also added documentation of default values.

I wish that I could tell you that this change is based on actual
concrete testing, but it is not. It is based on my own gut feeling
that overall, after #502, some things got worse and other things got
better, and that we should try and aim for the middle a bit more.  To
improve performance more, we should probably consider just reducing
the number of biters that we are willing to spawn in waves
(i.e. essentially have more boss biters and fewer normal biters).