Fix a race condition where players could volunteer to be captain as the game is starting
This would cause these players to be listed as a "captain" in the
captains list even though they were not the captain of the team.

Thank you outstanding/thesoldier for the bug report

I also fixed a potential one on someone joining the game just as it
starts, which could have made them show up in the second round of
picking, but not the first round, which would be strange.