Clean up errors in log (#528)
* Partially revert pathfinder settings change (#502)

This is primarily motivated because I believe that the high value of
short_cache_size was drastically hurting performance in some
situations, based on
https://forums.factorio.com/viewtopic.php?p=610324#p610324

So, the main thing that I wanted to do here is to move things closer
to their default values, and closer to the values that they were
before #502.  Note that I am leaving the lower-pathfinder-resolution,
as I assume that that is still a big speedup for the pathfinder in
general (even though the downsides are substantial).

I also added documentation of default values.

I wish that I could tell you that this change is based on actual
concrete testing, but it is not. It is based on my own gut feeling
that overall, after #502, some things got worse and other things got
better, and that we should try and aim for the middle a bit more.  To
improve performance more, we should probably consider just reducing
the number of biters that we are willing to spawn in waves
(i.e. essentially have more boss biters and fewer normal biters).

* Find end-of-game bug with biter targeting

Without this change, we can end up with log entries like the following:

    .../temp/currently-playing/utils/event_core.lua:30: Error caught: Given entity doesn't exist anymore.
    .../temp/currently-playing/utils/event_core.lua:32: stack traceback:
    .../temp/currently-playing/utils/event_core.lua:32: in function <...ctorio1.1.42/temp/currently-playing/utils/event_core.lua:29>
    [C]: in function 'set_multi_command'
    .../temp/currently-playing/maps/biter_battles_v2/ai.lua:51: in function '?'
    .../temp/currently-playing/maps/biter_battles_v2/main.lua:408: in function <...42/temp/currently-playing/maps/biter_battles_v2/main.lua:388>
    [C]: in function 'xpcall'
    .../temp/currently-playing/utils/event_core.lua:49: in function 'call_handlers'
    .../temp/currently-playing/utils/event_core.lua:61: in function <...ctorio1.1.42/temp/currently-playing/utils/event_core.lua:56>

or

    .../temp/currently-playing/utils/event_core.lua:30: Error caught: Given entity doesn't exist anymore.
    .../temp/currently-playing/utils/event_core.lua:32: stack traceback:
    .../temp/currently-playing/utils/event_core.lua:32: in function <...ctorio1.1.42/temp/currently-playing/utils/event_core.lua:29>
    [C]: in function 'set_command'
    ...temp/currently-playing/maps/biter_battles_v2/ai_strikes.lua:140: in function 'assassinate'
    ...temp/currently-playing/maps/biter_battles_v2/ai_strikes.lua:184: in function 'step'
    ...temp/currently-playing/maps/biter_battles_v2/main.lua:316: in function <...42/temp/currently-playing/maps/biter_battles_v2/main.lua:315>
    [C]: in function 'xpcall'
    .../temp/currently-playing/utils/event_core.lua:49: in function 'call_handlers'
    .../temp/currently-playing/utils/event_core.lua:61: in function <...ctorio1.1.42/temp/currently-playing/utils/event_core.lua:56>

I am just trying to clean up all errors from the log.

* More error cleanup from logs

...temp/currently-playing/maps/biter_battles_v2/feeding.lua:315: Error in calc_send: "...y-playing/maps/biter_battles_v2/feeding_calculations.lua:154: attempt to perform arithmetic on local 'flask_count' (a nil value)"  parameter: "force=north color=white count=15k"

* and more error log cleanup

I worry that this one might be a bug in some capacity, perhaps someone
could exploit this to bring in copy history somehow?  Sounds weird to
me though.

.../temp/currently-playing/utils/event_core.lua:30: Error caught: ....42/temp/currently-playing/maps/biter_battles_v2/gui.lua:47: attempt to index field 'cursor_stack' (
a nil value)
62633.288 Script @/opt/factorio1.1.42/temp/currently-playing/utils/event_core.lua:32: stack traceback:
        ...ctorio1.1.42/temp/currently-playing/utils/event_core.lua:32: in function '__index'
        ....42/temp/currently-playing/maps/biter_battles_v2/gui.lua:47: in function 'clear_copy_history'
        ...42/temp/currently-playing/maps/biter_battles_v2/main.lua:39: in function <...42/temp/currently-playing/maps/biter_battles_v2/main.lua:33>
        [C]: in function 'xpcall'
        ...ctorio1.1.42/temp/currently-playing/utils/event_core.lua:49: in function 'call_handlers'
        ...ctorio1.1.42/temp/currently-playing/utils/event_core.lua:61: in function <...ctorio1.1.42/temp/currently-playing/utils/event_core.lua:56>