captains join info: make popup appear when joining more often
This makes the UI slightly more confusing for referees, because both
windows can be open at the same time, and on top of one another. But I
still think that this is overall an improvement.

Without this change, if the server had restarted recently, such that
it would show the rules when you join, then the "Join Info" window
would not be open when joining a captains game.

It would be nice if the initial window positions could be chosen a bit
better so that they didn't start 100% overlapping for the referee, but
it is probably ok as-is.