Always show evo/threat
This also removes the "lobby" functionality, which apparently only
really was used when the server was fully restarted after each game.

This also removes the text that used to appear before joining a team
of "Defend your rocket silo!" and "Feed the enemy team's biters to
gain advantage!".  We will have to rely on the "?" button for that
context I think, or we can add the text back in if people think that
it is valuable.

This also simplifies code substantially by drawing the top-left "main"
GUI using the same codepath whether you have been assigned a team or
not.