Main gui: make buttons much more responsive
Specifically, before this change, the buttons would be destroyed +
recreated every few seconds. If this destroy+recreate happened between
the click of a button, and the actual tick that the click is
recognized by the server, then the button click would have been lost
before, and now it will not be. Thus, for users with high latency
connections, it was very unreliable to click "join team" or "send
science" buttons. After this change, the buttons should no longer be
unreliable.

Note that this means that we potentially need to deal with
double-clicks of buttons like the join button. Before, this was
impossible, as the button was destroyed when it was clicked, and so
the game engine would ignore any subsequent clicks. Now, we only hide
the button (rather than destroy it), and so the game will still
potentially process multiple clicks. Thus we change join_team() to
no-op if the player has already joined a team.