captains: change initial picking to be 1-2-2-2-2... always
Note that the picking system is more accurately described as "captains
can continue to pick players until their team has more players than
the other team". So if there is one group of 3 that is picked first,
then the picking would go 3-4-2-2-2-...

Note that after the initial picking phase, it is always just
captains-alternate-picking.