captains: allow players to un-volunteer to play/captain
I'm not 100% sure if this is a good idea. But, often people are more
willing to try things if it feels like less of a commitment, i.e. if
they know that they can back out later on.  We can roll this back
later on if it causes problems.

I also added a log statement for when the game starts, and made it
easier to test the captains game code when playtime isn't hooked up.