captains: fix some weird conditions with "I do not want to play/captain" (#553)
There are two fixes here. One is straightforward, where it was odd
that if you volunteered to be captain, but then said "actually, I
don't want to play", you would remain as a captain volunteer.

The other is that I think that on a high latency connection, it would
have been possible to do weird things (like volunteer to be captain
*after* you had stated that you don't want to play) to end up in the
same state.

Specifically, because we only change the visibility of these buttons,
without actually destroying/recreating them, it is possible for
network latency to cause clicks to appear even when the button would
otherwise be disabled. Perhaps these could be fixed by checking for
element.enabled in all cases, but I'm not sure.

Anyhow, I think that this should fix the issue that sticklord produced
where they were volunteering to be captain but not in the list of
players wanting to play.