End game stats UI (/teamstats command) (#559)
* refactor some safety logic to use for more commands

As I discovered with /inventory-costs, it is easy to accidentally
crash the server due to a logic bug.  This helps avoid that.

* inventory-costs: use new safe_wrap function for safety

* Add end-of-game UI, also sometimes available with /teamstats command

This displays many stats that are useful to understand how the game
went. By default, it is only available after the game ends. You can
also see stats for the previous game with "/teamstats prev"

This change was enabled by proxc, who built awesome mock-ups of what
the information should look like.

This should have almost no performance overhead, as we are just
looking at production statistics once per 16 seconds, and running a
little bit of extra code in the on_entity_died handler.