/close-difficulty-vote command: add optional param to force difficulty to a given value (#562)
This gives a simpler option for admins to force the difficulty. Before
this, it was necessary to first go Fish -> Config ->
Admin-only-difficulty-voting-on, then run /difficulty-revote, then
vote, then run /difficulty-close-vote.  Now you can just run one
command.

This also fixes the difficulty-name/value listed in '/teamstats prev'
which was previously broken, and improves some formating of long-team
names in the /teamstats view.