Add event priority (#588)
Added event priorities to event.lua. Event handlers with the highest priority will now always be called first, priorities can be any number and defaults to 0. Event handlers with the same priority will always be called in order of insertion (or in order of event.add calls).
Simplification of some part of removable functions logic.
Use of event priority to fix this problem.

Co-authored-by: Cliff Frey <clifffrey@gmail.com>