captains games: add 'community picked teams' mode
In this mode, players of the game itself must select a good player
picking order in order to create fair teams.  Every player that wants
to play in the game is given the opportunity to build a
stack-rank/pick-order of all players that want to play. Once at least
5 players have done so (and at least 30% of all players have done so)
it is possible for the referee to start the game.

The algorithm for picking is the same 1-2-2-2-... picking order. For
each pick, we treat the lists as a ranked-choice vote.  Basically look
at everyone's first choice. If one candidate has >50% of the vote,
then they are picked. Otherwise, we remove the candidate with the
fewest votes from being eligible to be picked this round and start
over. Eventually, it is guaranteed that one player will have >50% of
the vote. They are then picked for a team, and then we repeat until
all players have been picked.

Given that all players are picking at the same time, in parallel, it
should be difficult for players to somehow conspire to impact which
players will be on their team.

There are some additional (mostly unrelated) changes in here as well.

1) The tournament UI is now always in "collapsed" form. I think that
this form is simpler/easier, as it involves fewer nested scrollbars. I
also think that it makes it easier for players to figure out how to
join, as "escape" won't automatically close the join window.

2) I changed "I want to be a PLAYER" to "I want to play"

3) I removed the old "balanced random teams" game mode, as it was
rarely used, and was difficult to actually get balanced teams out of
it.

4) change text around the extra-player-info to make it clear that it
might be more-widely-visible than just captains seeing it.