community-pick: have a volunteer-to-be-captain phase, and allow 'no captain'
Captains games now support having a team be "captainless" (the actual
captain variable is nil). In this case, all players are trusted to
send science. When latejoiners need to be picked, they will be picked
by a random player on the team chosen among the top 30% of players
sorted by pick-order who are currently online.  In all captains games
(whether they are community picked or not), this can be arranged with
commands like `/replaceCaptainNorth !nobody!`.

Now, after a 'community-pick' game assigns people to teams, for each
team, the players in the top 30% of the "pick order" are prompted to
ask if they want to be a captain. They can answer one of Yes/Maybe/No
(if they don't answer, it is the same as answering "no").  After 30
seconds, if people answered "Yes", then one of them is chosen at
random to be captain. Else if people answered "Maybe", then one of
them is chosen to be captain, else the team is captainless.

Also, 'community-pick' games now allow people to say "I don't want to
play" before the game starts. All players that have already put that
player in a pick order just stay as they were (and if the player
changes their mind to play *again*, no more work is required by that
user). Thus, even if someone keeps changing their mind about playing,
it will not impact the game's start very much.

I also slightly reduced the maximum height of the "Join info" screen,
as people seemed to complain about its size.

Also, 'community-pick' games can now be started with as few as 2
people who have confirmed their ranks. This is primarily to make
testing easier for me.

I also fixed all (!) flagged LSP errors in captain.lua. This required
a bunch of annotations like `--[[@as string]]` and some amount of code
restructuring so that the type checker could validate that things were
safe.  I just found that the type checker was actually finding a bunch
of real issues (like passing the wrong number of arguments to
functions), so it felt worthwhile to me to clean it up.