Minimal 2.0.x support (#625)
* 2.0: rename global -> storage

* 2.0: adapt debug.getinfo(n, 'S') source field

This introspective function no longer returns absolute path to scenario.
Instead of '/home/$USER/.../currently-playing/..' it's '@__level__/..'

* 2.0: ai_targets: adapt for on_object_destroyed

* 2.0: require: fix paackage paths in require()

Lua vm seems to now store loaded packages in different format, so patch
those on the fly to avoid change of standard semantics.
(e.g. utils.event -> __level__/utils/event.lua)

* 2.0: remove use of heading-3 font

Seems that support of this font was dropped on purpose or accident.
Replace it with similar default-semibold to get rid of this error:
 Error Util.cpp:81: Missing specification of font with name: "heading-3"

* 2.0: drop usage of research_queue_enabled

It's now enabled by default.

* 2.0: move from LuaGameScript::active_mods to LuaBoostrap::active_mods

* 2.0: drop the -layer suffix from collision masks

* 2.0: antigrief: check event.player_index

In 2.0 this argument is optional, if permission group was modified by mod.

* 2.0: bunch of adaptations to GUI

A lot of sprites were removed, I tried to find similar looking, but
sometimes it was not possible at all, when for example _hover or _black
version of some sprite was removed.

Same thing happened to a lot of very specific GUI themes, in those cases
I just replaced it with more general forms.

* 2.0: gui: add explicit bool checks for 'storage.bb_game_won_by_team'

For some reason we cannot use nil as bool anymore. Was this always the
case or am I crazy?

* run stylua

---------

Co-authored-by: Cliff Frey <clifffrey@gmail.com>