Adjust nests value by evolution (#633)
Biter nests health increases with evo (see FFF-427). This adjusts the score obtained by killing spawners according to the new approximated function of:
    y = 350 + 3150 * (x^2.25)
resulting in
    factor = 1 + 9 * (evolution^2.25)

which describes the factor of nest_health / base_prototype_health in function of evo(x). Note that this only applies up to 100% evo, after that, the threat reduction for killing (and health value of) the biter nests does not change.