terrain: remove expensive tile replacement
Although in a lot of cases get_replacement_tile() is successful on first try
the amount of checks performed explodes on a surface with starting lake
generated in spawn area. In my testing with seed 20690612 which has
small lake, we gain ~20% speedup (1s -> 800ms) by getting rid of this
function and defaulting to constant value. Test was performed using i9-12900KF.
This function contributes to some people dropping from the game during
map reset as they cannot catch-up.

Getting rid of this function itself is barely noticable, because tile we
replace is going under concrete which is very rarely removed anyway.