gui: spectate: ignore LuaPlayer::spectator
Don't use LuaPlayer::spectator property to judge whether player is in
fact spectator or not. This property has nothing to do with force
assignment and can be manipulated from player setting in comfy panel
at will. This can lead to a situation where player without a character
is moved to spectator island and spectator initialization code is not
executed for them.