Update ai.lua
"We buff Yellow and White Sci by 25%. So Mutagen Value on ITYTD is the same as before, so ITYTD will remain White-Sci Meta. We think White Sci sendings vs. Defence is on a balanced level. Therefor we like to delay that Diff%-Ramp on all Levels by 30 Minutes, because we think with this 25% boost on White, these will hit too strong otherwise."